Check entry points
Adds a set of basic checks for the `build`, `browserify`, `run`, and
`test` commands that the entry point module is appropriate as an entry
point, and gives better errors if there's something wrong with it (e.g.
the entry point module does not exist, or does not export main, or
exports main but it has the wrong type).

refs purescript/purescript#2086