Stop attempting to register on Bower registry during publish
Instead of automatically registering packages on the Bower registry
during publish, we now check to see whether a package is registered in
purescript/registry, and if it is not, throw an error and ask the user
to register manually. This should allow new packages to be published as
the bower registry is no longer accepting new package registrations.

See also https://github.com/purescript/package-sets/pull/577/files.

There are no tests, but I have tested this manually and I'm fairly
confident it works.